// src/lib/theme/types.ts
export type ButtonVariant =
  | "primary"
  | "secondary"
  | "outline"
  | "ghost"
  | "danger"
  | "gradient";
export type ButtonSize = "sm" | "md" | "lg";